libs/Radx/NcfRadxFile - adding proper support for flag_values and flag_meanings

+++ codebase/libs/Radx/src/Nexrad/NexradRadxFile.cc
@@ -2751,7 +2751,7 @@ int NexradRadxFile::_writeMetaDataHdrs(const RadxVol &vol)
   if(_adap.beamwidth == 0)
   {
     const string site_name = vol.getSiteName();
-    for (int i=0; i<site_name.size(); i++) {
+    for (size_t i=0; i<site_name.size(); i++) {
       _adap.site_name[i] = site_name[i];
     }
     _adap.antenna_gain = vol.getRadarAntennaGainDbH();
